Parallelize daily solutions
This brings down the total runtime from ~700ms to ~240ms on my machine,
as measured by:

    cargo build -r && hyperfine target/release/aoc

Note that now the total runtime is bounded by the slowest daily solution
(day 6). Rayon parallelization could be used inside daily solutions, but
i think the gains are not worth it. They are fast enough, and i'd prefer
to keep them as simple as possible.