Use std threads instead of Rayon
Since we're only doing coarse-grained parallelization of daily
solutions, Rayon doesn't add much. Ditch it in favour of std threads.
This doesn't affect runtime noticeably (it's just 25 threads), and has
the nice side effect of daily answers being printed as soon as they are
ready instead of waiting for all of them to finish before printing.